create kodawari.kafka package
