+++ py/lib/kodawari.kafka/pyproject.toml
@@ -1,0 +1,25 @@
+[tool.poetry]
+name = "kodawari.kafka"
+version = "0.1.0"
+description = "A package for kodawari kafka utilities"
+authors = ["Zachary Groff <groff.zackm@gmail.com>"]
+readme = "README.md"
+packages = [{include = "kafka"}]
+
+[tool.poetry.dependencies]
+python = "^3.11"
+confluent-kafka = "^2.0.2"
+kodawari-models = {path = "../../lib/kodawari.models", develop = true}
+pydantic = "^1.10.4"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.2.1"
+pyright = "^1.1.288"
+isort = "^5.11.4"
+ssort = "^0.11.6"
+black = "^22.12.0"
+darglint = "^1.8.1"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
